Rename getUpdate for consistency
